feat(chore): packages / aliases rearrange
