+++ packages/serverless-utils/project.json
@@ -6,14 +6,20 @@
   "targets": {
     "lint": {
       "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
+      "outputs": [
+        "{options.outputFile}"
+      ],
       "options": {
-        "lintFilePatterns": ["packages/serverless-utils/**/*.ts"]
+        "lintFilePatterns": [
+          "packages/serverless-utils/**/*.ts"
+        ]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "outputs": [
+        "{workspaceRoot}/coverage/{projectRoot}"
+      ],
       "options": {
         "jestConfig": "packages/serverless-utils/jest.config.ts",
         "passWithNoTests": true
@@ -26,5 +32,7 @@
       }
     }
   },
-  "tags": ["lib"]
+  "tags": [
+    "lib"
+  ]
 }
