--- packages/serverless-utils/src/index.ts
@@ -1,11 +1,0 @@
-// Types
-
-// Constants
-export { ResponseHeader } from './lib/headers';
-
-// Functions
-export { buildLambdaHandlerPath } from './lib/serverless-utils';
-export { formatJSONSuccessResponse } from './lib/serverless-utils';
-export { formatErrorResponse } from './lib/serverless-utils';
-export { mapItemsById, generateUUID } from './lib/serverless-utils';
-export { ValidationError, validationError } from './lib/serverless-utils';
