+++ stacks/products-service/README.md
@@ -6,6 +6,8 @@ This stack was generated with [Nx](https://nx.dev).
 
 Run `nx test products-service` to execute the unit tests via [Jest](https://jestjs.io).
 
-Run `nx build products-service` to package code via [Serverless Packaging](https://www.serverless.com/framework/docs/providers/aws/guide/packaging).
+Run `nx build products-service` to package code
+via [Serverless Packaging](https://www.serverless.com/framework/docs/providers/aws/guide/packaging).
 
-Run `nx deploy products-service` to deploy code to AWS via [Serverless Deploying](https://www.serverless.com/framework/docs/providers/aws/guide/deploying).
+Run `nx deploy products-service` to deploy code to AWS
+via [Serverless Deploying](https://www.serverless.com/framework/docs/providers/aws/guide/deploying).
