+++ stacks/products-service/src/functions/products-create/handler.ts
@@ -1,49 +1,12 @@
-import {
-  formatErrorResponse,
-  formatJSONSuccessResponse,
-  generateUUID,
-  ValidationError,
-} from '@aws-practitioner-training/serverless-utils';
-import { DynamoDBDocumentClient } from '@aws-sdk/lib-dynamodb';
-import { DynamoDBClient, ExecuteTransactionCommand } from '@aws-sdk/client-dynamodb';
-import { assertProductIsValid, AvailableProduct } from '../lib';
-
+import { formatErrorResponse, formatJSONSuccessResponse } from '@helpers/common';
+import { assertProductIsValid, ValidationError } from '@helpers/validation';
 import { APIGatewayProxyEvent } from 'aws-lambda';
-import { isString } from '@powwow-js/core';
-
-const ProductsTableName = process.env.ProductsTableName;
-const StocksTableName = process.env.StocksTableName;
-const AwsRegion = process.env.AwsRegion;
-
-const dynamo = DynamoDBDocumentClient.from(new DynamoDBClient({ region: AwsRegion }));
-
-export const createAvailableProduct = async (product: AvailableProduct) => {
-  const productId = generateUUID();
-  const transactionalInsert = new ExecuteTransactionCommand({
-    TransactStatements: [
-      {
-        Statement: `INSERT INTO "${ProductsTableName}" VALUE {'id':?, 'title':?, 'description':?, 'price':?}`,
-        Parameters: [
-          { S: productId },
-          { S: product.title },
-          { S: product.description },
-          { N: product.price.toString() },
-        ],
-      },
-      {
-        Statement: `INSERT INTO "${StocksTableName}" VALUE {'productId':?, 'count':?}`,
-        Parameters: [{ S: productId }, { N: product.count?.toString() ?? '0' }],
-      },
-    ],
-  });
-  const output = await dynamo.send(transactionalInsert);
-  return { productId, output };
-};
+import { createAvailableProduct } from '@helpers/db-client';
 
 export const main = async (event: APIGatewayProxyEvent) => {
   try {
     console.log('~~~~~ Payload: ', event.body);
-    const product = isString(event.body) ? JSON.parse(event.body) : event.body;
+    const product = JSON.parse(event.body);
     assertProductIsValid(product);
     const { productId, output } = await createAvailableProduct(product);
     console.log('~~~~~ DB::created: ', `id:${productId}`, output);
