+++ stacks/products-service/src/functions/products-list/handler.spec.ts
@@ -1,3 +1,4 @@
+import * as dbClient from '@helpers/db-client';
 import * as productsListHandler from './handler';
 
 describe('products-list tests', () => {
@@ -12,7 +13,7 @@ describe('products-list tests', () => {
   });
 
   it('should find product by id', async () => {
-    const spyOn = jest.spyOn(productsListHandler, 'getAvailableProductItems').mockImplementation(() =>
+    const spyOn = jest.spyOn(dbClient, 'getAvailableProductItems').mockImplementation(() =>
       Promise.resolve([
         { id: '1', title: 'Test Product 1', count: 1, description: '11', price: 10 },
         { id: '2', title: 'Test Product 2', count: 2, description: '22', price: 22 },
