+++ stacks/products-service/src/functions/products-list/handler.ts
@@ -1,34 +1,6 @@
-import {
-  formatErrorResponse,
-  formatJSONSuccessResponse,
-  mapItemsById,
-} from '@aws-practitioner-training/serverless-utils';
-import { APIGatewayProxyEvent } from 'aws-lambda';
-import { DynamoDBDocumentClient, ScanCommand } from '@aws-sdk/lib-dynamodb';
-import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
-import { ProductDbItem, StockDbItem } from '../lib';
-
-const ProductsTableName = process.env.ProductsTableName;
-const StocksTableName = process.env.StocksTableName;
-const AwsRegion = process.env.AwsRegion;
-
-const dynamo = DynamoDBDocumentClient.from(new DynamoDBClient({ region: AwsRegion }));
-
-export const getAvailableProductItems = async () => {
-  const [products, stocks] = await Promise.all([
-    dynamo.send(new ScanCommand({ TableName: ProductsTableName })),
-    dynamo.send(new ScanCommand({ TableName: StocksTableName })),
-  ]);
-
-  const stocksMap = mapItemsById<StockDbItem>('productId', <StockDbItem[]>stocks.Items);
-
-  return products.Items.map((product: ProductDbItem) => {
-    return {
-      ...product,
-      count: stocksMap.get(product.id)?.count ?? 0,
-    };
-  });
-};
+import { formatErrorResponse, formatJSONSuccessResponse } from "@helpers/common";
+import { APIGatewayProxyEvent } from "aws-lambda";
+import { getAvailableProductItems } from "@helpers/db-client";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const main = async (_event: APIGatewayProxyEvent) => {
