+++ tsconfig.base.json
@@ -12,16 +12,30 @@
     "module": "commonjs",
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-    "lib": ["es2017", "dom"],
+    "lib": [
+      "es2017",
+      "dom"
+    ],
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@aws-practitioner-training/serverless-utils": ["packages/serverless-utils/src/index.ts"]
+      "@helpers/common": [
+        "packages/serverless-utils/src/lib/common/index.ts"
+      ],
+      "@helpers/db-client": [
+        "packages/serverless-utils/src/lib/db-client/index.ts"
+      ],
+      "@helpers/validation": [
+        "packages/serverless-utils/src/lib/validation/index.ts"
+      ]
     },
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "removeComments": true
   },
-  "exclude": ["node_modules", "tmp"]
+  "exclude": [
+    "node_modules",
+    "tmp"
+  ]
 }
